made auth mandatory for admin priviledges

+++ src/api/category.ts
@@ -1,3 +1,4 @@
+import fetchUser from '../middleware/fetchUser';
 import express from 'express';
 import prisma from '../instance/prisma-instance';
 import bodyParser from 'body-parser';
@@ -19,39 +20,77 @@ router.get('/getCategory/:id', async (req, res) => {
   res.send(category);
 });
 
-router.post('/createCategory', async (req, res) => {
-  const createCategory = await prisma.category.create({
-    data: {
-      name: req.body.name,
+router.post('/createCategory', fetchUser, async (req: any, res) => {
+  const userId = req.user.id;
+  const user = await prisma.user.findFirst({
+    where: {
+      id: parseInt(userId),
     },
   });
-  res.send(
-    `Category with name: ${createCategory.name} has been added successfully!`
-  );
+  if (user?.roles == 'ADMIN') {
+    const createCategory = await prisma.category.create({
+      data: {
+        name: req.body.name,
+      },
+    });
+    res.send(
+      `Category with name: ${createCategory.name} has been added successfully!`
+    );
+  } else {
+    res.send(
+      'This request needs Admin level priviledges. Please try logging in with admin credentials.'
+    );
+  }
 });
 
-router.patch('/updateCategory/:id', async (req, res) => {
-  const updateCategory = await prisma.category.update({
+router.patch('/updateCategory/:id', fetchUser, async (req: any, res) => {
+  const userId = req.user.id;
+  const user = await prisma.user.findFirst({
     where: {
-      id: parseInt(req.params.id),
-    },
-    data: {
-      name: req.body.name,
+      id: parseInt(userId),
     },
   });
-  console.log(updateCategory);
-  res.send(`Category with id: ${req.params.id} has been updated successfully!`);
+  if (user?.roles == 'ADMIN') {
+    const updateCategory = await prisma.category.update({
+      where: {
+        id: parseInt(req.params.id),
+      },
+      data: {
+        name: req.body.name,
+      },
+    });
+    console.log(updateCategory);
+    res.send(
+      `Category with id: ${req.params.id} has been updated successfully!`
+    );
+  } else {
+    res.send(
+      'This request needs Admin level priviledges. Please try logging in with admin credentials.'
+    );
+  }
 });
 
-router.delete('/deleteCategory/:id', async (req, res) => {
-  const deleteCategory = await prisma.category.delete({
+router.delete('/deleteCategory/:id', fetchUser, async (req: any, res) => {
+  const userId = req.user.id;
+  const user = await prisma.user.findFirst({
     where: {
-      id: parseInt(req.params.id),
+      id: parseInt(userId),
     },
   });
-  res.send(
-    `Category with name: ${deleteCategory.name} has been deleted successfully!`
-  );
+  if (user?.roles == 'ADMIN') {
+    const deleteCategory = await prisma.category.delete({
+      where: {
+        id: parseInt(req.params.id),
+      },
+    });
+    res.send(
+      `Category with name: ${deleteCategory.name} has been deleted successfully!`
+    );
+  } else {
+    res.send(
+      'This request needs Admin level priviledges. Please try logging in with admin credentials.'
+    );
+  }
 });
 
 export default router;
